url/urlsearchparams-foreach.html

<!doctype html>
<meta charset="utf-8">
<link rel="help" href="https://url.spec.whatwg.org/#dom-urlsearchparams-has">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    var params = new URLSearchParams('a=1&b=2&c=3');
    var keys = [];
    var values = [];
    params.forEach(function(value, key) {
        keys.push(key);
        values.push(value);
    });
    assert_array_equals(keys, ['a', 'b', 'c']);
    assert_array_equals(values, ['1', '2', '3']);
}, "ForEach Check");

test(function() {
    let a = new URL("http://a.b/c?a=1&b=2&c=3&d=4");
    let b = a.searchParams;
    var c = [];
    for (i of b) {
        a.search = "x=1&y=2&z=3";
        c.push(i);
    }
    assert_array_equals(c[0], ["a","1"]);
    assert_array_equals(c[1], ["y","2"]);
    assert_array_equals(c[2], ["z","3"]);
}, "For-of Check");

test(function() {
    let a = new URL("http://a.b/c");
    let b = a.searchParams;
    for (i of b) {
        assert_unreached(i);
    }
}, "empty");

test(function() {
    const a = new URL("http://a.b/c?a=1&b=2&c=3");
    const b = a.searchParams;
    const c = [];
    for (const i of b) {
        if (i[0] === 'a') {
            b.delete('b');
        }
        c.push(i);
    }
    assert_array_equals(c[0], ["a", "1"]);
    assert_array_equals(c[1], ["c", "3"]);
}, "delete next param during iteration");

test(function() {
    const a = new URL("http://a.b/c?a=1&b=2&c=3");
    const b = a.searchParams;
    const c = [];
    for (const i of b) {
        if (i[0] === 'a') {
            b.delete('a');
            // 'b=2' is now in the first slot, so the next iteration will see 'c=3'.
        } else {
            c.push(i);
        }
    }
    assert_array_equals(c[0], ["c", "3"]);
}, "delete current param during iteration");

test(function() {
    const a = new URL("http://a.b/c?a=1&b=2&c=3");
    const b = a.searchParams;
    const c = [];
    for (const i of b) {
        c.push(i[0]);
        b.delete(i[0]);
    }
    assert_array_equals(c, ['a', 'c'], 'b should not have been seen by the loop');
    assert_equals(String(b), 'b=2', 'b should remain');
}, "delete every param seen during iteration");
</script>
